# Conflicts:
#	src/knn_classify_tinycifar10.py

+++ src/ImageVectorizer.py
@@ -1,5 +1,5 @@
 from FeatureVectorDataset import FeatureVectorDataset
-
+import numpy as np
 class ImageVectorizer(FeatureVectorDataset):
     # Wraps an image dataset and exposes its contents.
     # Samples obtained using sample() are returned as 1D feature vectors.
@@ -10,7 +10,7 @@ class ImageVectorizer(FeatureVectorDataset):
         # Ctor.  dataset is the dataset to wrap (type ImageDataset).
 
     def devectorize(self, fvec):
-        return fvec.reshape((32,32,3))
+        return fvec.astype('uint8', copy=False).reshape((32,32,3))
         # Convert a feature vector fvec obtained using sample()
         # back to an image and return the converted version.
